test(bench): decouple surface tests from prebuild outputs
