+++ packages/remnic-cli/src/index.ts
@@ -128,11 +128,22 @@ import {
   type BenchmarkDefinition,
 } from "@remnic/bench";
 import { firstSuccessfulCandidate, firstSuccessfulResult } from "./service-candidates.js";
+import {
+  type BenchAction,
+  type ParsedBenchArgs,
+  parseBenchActionArgs,
+  parseBenchArgs,
+} from "./bench-args.js";
 export { hasFlag, resolveFlag, stripResolveFlags, TAXONOMY_RESOLVE_BOOLEAN_FLAGS } from "./cli-args.js";
 import { hasFlag, resolveFlag, stripResolveFlags, TAXONOMY_RESOLVE_BOOLEAN_FLAGS } from "./cli-args.js";
 import { parseConnectorConfig, stripConfigArgv } from "./parse-connector-config.js";
 
 export { parseConnectorConfig, stripConfigArgv };
+export {
+  type BenchAction,
+  type ParsedBenchArgs,
+  parseBenchArgs,
+} from "./bench-args.js";
 
 // ── Host-specific publisher registrations ───────────────────────────────────
 // Publisher classes live in @remnic/core, but wiring them into the registry
@@ -173,22 +184,11 @@ type CommandName =
 type DaemonAction = "start" | "stop" | "restart" | "install" | "uninstall" | "status";
 type TokenAction = "generate" | "list" | "revoke";
 type ReviewAction = "approve" | "dismiss" | "flag";
-export type BenchAction = "help" | "list" | "run" | "check" | "report";
-
 export interface BenchCatalogEntry {
   id: string;
   title: string;
   category: "agentic" | "retrieval" | "conversational";
   summary: string;
-}
-
-export interface ParsedBenchArgs {
-  action: BenchAction;
-  benchmarks: string[];
-  quick: boolean;
-  all: boolean;
-  json: boolean;
-  datasetDir?: string;
 }
 
 // ── Constants ────────────────────────────────────────────────────────────────
@@ -284,68 +284,6 @@ Examples:
   remnic bench run --quick longmemeval
   remnic bench run longmemeval --dataset-dir ~/datasets/longmemeval
   remnic benchmark run --quick longmemeval`;
-}
-
-function readBenchOptionValue(argv: string[], flag: string): string | undefined {
-  const index = argv.indexOf(flag);
-  if (index === -1) {
-    return undefined;
-  }
-
-  const value = argv[index + 1];
-  if (!value || value.startsWith("-")) {
-    throw new Error(`ERROR: ${flag} requires a value.`);
-  }
-
-  return value;
-}
-
-function collectBenchmarks(argv: string[]): string[] {
-  const benchmarks: string[] = [];
-  for (let index = 0; index < argv.length; index += 1) {
-    const arg = argv[index];
-    if (arg === "--dataset-dir") {
-      index += 1;
-      continue;
-    }
-    if (!arg.startsWith("-")) {
-      benchmarks.push(arg);
-    }
-  }
-  return benchmarks;
-}
-
-function parseBenchActionArgs(argv: string[]): {
-  action: BenchAction;
-  args: string[];
-} {
-  const [first, ...rest] = argv;
-  const action: BenchAction =
-    first === "list" || first === "run" || first === "check" || first === "report"
-      ? first
-      : first === undefined || first === "--help" || first === "-h"
-        ? "help"
-        : "run";
-
-  return {
-    action,
-    args: action === "run" && action !== first ? argv : rest,
-  };
-}
-
-export function parseBenchArgs(argv: string[]): ParsedBenchArgs {
-  const { action, args } = parseBenchActionArgs(argv);
-  const benchmarks = collectBenchmarks(args);
-  const datasetDir = readBenchOptionValue(args, "--dataset-dir");
-
-  return {
-    action,
-    benchmarks,
-    quick: args.includes("--quick"),
-    all: args.includes("--all"),
-    json: args.includes("--json"),
-    datasetDir: datasetDir ? path.resolve(expandTilde(datasetDir)) : undefined,
-  };
 }
 
 export function buildBenchRunnerArgs(
